+++ bin/geoipupdate_s.c
@@ -18,11 +18,6 @@ geoipupdate_s *geoipupdate_s_new(void) {
             "Unable to allocate memory for database directory path: %s\n",
             strerror(errno));
 
-    gu->proto = strdup("https");
-    exit_if(NULL == gu->proto,
-            "Unable to allocate memory for request protocol: %s\n",
-            strerror(errno));
-
     gu->host = strdup("updates.maxmind.com");
     exit_if(NULL == gu->host,
             "Unable to allocate memory for update host: %s\n",
@@ -60,7 +55,6 @@ void geoipupdate_s_delete(geoipupdate_s *gu) {
         edition_delete_all(gu);
         free(gu->license_file);
         free(gu->database_dir);
-        free(gu->proto);
         free(gu->proxy);
         free(gu->proxy_user_password);
         free(gu->lock_file);